Crowdfunding y merge

+++ resources/views/AdminPageView/AdminPageViewService/AdminPageViewServiceShortlink/adminPageviewserviceshortlink.blade.php
@@ -1,11 +1,14 @@
 @extends('AdminPageView.AdminPageViewTemplate.bodyadminpage')
 
 @section('content')
+@php
+    use App\Http\Controllers\LibraryFunctionController as LFC;
+@endphp
 <div class="container-fluid pt-4 px-4">
     <div class="row p-2 bg-light rounded  justify-content-center mx-0">
         <div class="row">
             <h1 class="my-2 fs-4 fw-bold text-center">LDK Syahid URL Shortener</h1>
-            <form action="{{ route('url.shorten') }}" method="POST" class="my-2">
+            <form action="{{ route('admin.service.shortlink.shorten') }}" method="POST" class="my-2">
                 @csrf
                 @method('POST')
                 <div class="input-group mb-3">
@@ -40,7 +43,11 @@
                         <td align="center">{{ $item->visits->count() }}</td>
                         <td align="center">
                             <button type="button" class="btn btn-sm btn-primary mb-1" data-bs-toggle="modal" data-bs-target="#exampleModal-{{ $key }}"><i class="fa fa-edit"></i></button>
-                            <button type="button" onclick="deleteConfirmationShortlink({{$item->id}})" id="delete-shortlink" class="btn btn-sm btn-primary mb-1"><i class="fa fa-trash"></i></button>
+                            @if (LFC::getRoleName(auth()->user()->getRoleNames()) == 'Superadmin')
+                                <button type="button" onclick="deleteConfirmationShortlink({{$item->id}})" id="delete-shortlink" class="btn btn-sm btn-primary mb-1"><i class="fa fa-trash"></i></button>
+                            @else
+
+                            @endif
                         </td>
                     </tr>
                     <!-- Modal -->
@@ -52,7 +59,7 @@
                                     <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
                                 </div>
                                 <div class="modal-body">
-                                    <form action="{{ route('update', $item->id) }}" method="POST">
+                                    <form action="{{ route('admin.service.shortlink.update', $item->id) }}" method="POST">
                                         @csrf
                                         <div class="mb-3">
                                             <label for="key" class="form-label">URL Key</label>
@@ -81,7 +88,7 @@
 @endsection
 
 @section('scripts')
-<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p" crossorigin="anonymous"></script>
+{{-- <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p" crossorigin="anonymous"></script> --}}
 <script>
 // ===== START CRUD ARTICLE =====
 // ini untuk konfirmasi delete
